watcher keep alive even get 304 not modified (#36)
Hi all! I'm really happy to see this project. I'm working on some python project and it use central dogma so this is what I really want! I've test it in my project and I've found the repository watcher is dead after 304 response.

I think it because here. new_latest revision is None when 304 returned so schedule_watch is not triggered. So I moved schedule_watch out from if statement. Could you please check if it makes sense to you? I worried about stackeoverflow but, focused on keeping alive watcher for now.

Co-authored-by: Seunggon Kim <crosien@gmail.com>